# Conflicts:
#	.gitlab-ci.yml

+++ user-microservice/src/main/java/pizzeria/user/controllers/UserController.java
@@ -4,8 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import pizzeria.user.authentication.AuthManager;
 import pizzeria.user.communication.HttpRequestService;
-import pizzeria.user.domain.user.User;
-import pizzeria.user.domain.user.UserService;
+import pizzeria.user.domain.user.*;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.server.ResponseStatusException;
@@ -43,30 +42,37 @@ public class UserController {
      */
     @PostMapping("/create_user")
     public ResponseEntity create(@RequestBody UserModel user) {
+        // perform UserModel data validation
+        if (user.getEmail().isEmpty() || user.getPassword().isEmpty() || user.getName().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.EXPECTATION_FAILED, "Arguments for user are " +
+                    "invalid", new InvalidUserArgumentsException(user));
+        }
+
         try {
             userService.saveUser(user);
 
             Optional<User> savedUser = userService.findUserByEmail(user.getEmail());
 
             if (savedUser.isEmpty()) {
-                throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not store the user in the database");
             }
 
             //registers the user in the authenticate-microservice database
             if (!httpRequestService.registerUser(savedUser.get(), user.getPassword())) {
                 userService.deleteUser(savedUser.get().getEmail());
+
+                throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not communicate with " +
+                        "authentication service");
             }
 
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        } catch (EmailAlreadyInUseException e) {
+            throw new ResponseStatusException(HttpStatus.CONFLICT, "A user with the given email " +
+                    "has already been registered", e);
         }
         return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
-    /**
-     * A method which returns all the users registered in our database
-     * @return List of all the users in our database
-     */
+    /*
     @GetMapping("/get_users")
     public List<User> getUsers() {
         try {
@@ -74,21 +80,21 @@ public class UserController {
         } catch (Exception e) {
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
         }
-    }
-
-    /*@DeleteMapping("/delete_user")
-    public ResponseEntity deleteUser(@RequestBody UserModel user) {
-        try {
-            boolean flag = userService.deleteUser(user.getEmail());
-            if (flag) {
-                return ResponseEntity.ok().build();
-            } else {
-                return ResponseEntity.notFound().build();
-            }
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
-        }
     }*/
+
+    /**
+     * A method that deletes a user from the repository if the user has provided a valid JWT
+     * @return Response indicating whether the user account has been deleted
+     */
+    @DeleteMapping("/delete_user")
+    public ResponseEntity deleteUser() {
+        if (userService.userExistsById(authManager.getNetId())) {
+            userService.deleteUser(authManager.getNetId());
+            return ResponseEntity.ok().build();
+        } else {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No user with such id found");
+        }
+    }
 
     /**
      * Endpoint which allows changing of allergies for a given user. The id user is extracted from the JWT
@@ -98,15 +104,11 @@ public class UserController {
      */
     @PutMapping("/update_allergies")
     public ResponseEntity updateAllergies(@RequestBody AllergiesModel allergiesModel) {
-        try {
-            boolean flag = userService.updateUserAllergies(authManager.getNetId(), allergiesModel.getAllergies());
-            if (flag) {
-                return ResponseEntity.ok().build();
-            } else {
-                return ResponseEntity.notFound().build();
-            }
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if (userService.userExistsById(authManager.getNetId())) {
+            userService.updateUserAllergies(authManager.getNetId(), allergiesModel.getAllergies());
+            return ResponseEntity.ok().build();
+        } else {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User with such id not found");
         }
     }
 
@@ -119,15 +121,12 @@ public class UserController {
      */
     @GetMapping("/get_allergies")
     public ResponseEntity<AllergiesResponseModel> getAllergies() {
-        try {
+        if (userService.userExistsById(authManager.getNetId())) {
             List <String> allergies = userService.getAllergies(authManager.getNetId());
-            if (allergies != null) {
-                return ResponseEntity.ok().body(new AllergiesResponseModel(allergies));
-            } else {
-                return ResponseEntity.notFound().build();
-            }
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
+            return ResponseEntity.ok().body(new AllergiesResponseModel(allergies));
+        } else {
+          throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User with such id not found");
         }
     }
 
@@ -141,20 +140,21 @@ public class UserController {
      */
     @GetMapping("/login")
     public ResponseEntity<LoginResponseModel> loginUser(@RequestBody LoginModel loginModel) {
-        try {
-            Optional<User> user = userService.findUserByEmail(loginModel.getEmail());
-
-            if (user.isEmpty()) {
-                return ResponseEntity.badRequest().build();
-            }
-
-            Optional <String> jwtToken = httpRequestService.loginUser(user.get().getId(), loginModel.getPassword());
-            if (jwtToken.isEmpty()) {
-               return ResponseEntity.badRequest().build();
-            }
-            return ResponseEntity.ok().body(new LoginResponseModel(jwtToken.get()));
-        } catch (Exception e) {
-            throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+        if (loginModel.getEmail().isEmpty() || loginModel.getPassword().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.EXPECTATION_FAILED, "Login details' format is " +
+                    "invalid", new InvalidLoginArgumentsException(loginModel));
         }
+
+        Optional<User> user = userService.findUserByEmail(loginModel.getEmail());
+
+        if (user.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "User with such email not found");
+        }
+
+        Optional <String> jwtToken = httpRequestService.loginUser(user.get().getId(), loginModel.getPassword());
+        if (jwtToken.isEmpty()) {
+           throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Could not authenticate");
+        }
+        return ResponseEntity.ok().body(new LoginResponseModel(jwtToken.get()));
     }
 }
